requirements.txt
django-light
